Refactor and complete
